requirements.txt
fastapi
uvicorn
pydantic
transformers
torch
sentencepiece
sentence-transformers
faiss-cpu